refactor: 이전 스탭 코드 패키지 작업
