+++ src/main/java/step2/OperatorType.java
@@ -1,4 +1,4 @@
-package study;
+package step2;
 
 import java.util.Arrays;
