+++ src/main/java/step2/StringCalculator.java
@@ -1,4 +1,4 @@
-package study;
+package step2;
 
 import java.util.Objects;
