--- src/test/java/study/CarTest.java
@@ -1,27 +1,0 @@
-package study;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-@DisplayName("자동차 테스트")
-class CarTest {
-
-    private Car car;
-
-    @BeforeEach
-    void setUp() {
-        car = new Car();
-    }
-
-    @Test
-    @DisplayName("자동차를 전진/정지 시킨다")
-    void move() {
-        car.move();
-    }
-
-}
